Add TestFrontend.java. Change string format for printing table. Change the way app is run to accommodate test file.
TestFrontend.java is from week of Feb.16-23. To accommodate pre-written tests, the way this app starts is changed to use a run method that takes in string arguments as well as a run method that takes in a backend object. Also added a message to help user know what mode they have entered.